Create RegArray and RangeArray classes to encode the reg and ranges properties
It's common to want to query nodes for their `reg` and `ranges`
proprties to extract the fields by address and size. This allows the
user to use `Node.get_reg()` and `Node.get_ranges()` to get the list of
tuples from these properties.

These helpers automatically group cells of the `reg` and `ranges`
properties by `#address-cells` and `#size-cells` and combine them into a
single integer.

Signed-off-by: Nathaniel Graff <nathaniel.graff@sifive.com>