Fix interpretation of the `ranges` property
I was wrong about my interpretation of the `ranges` property, and so the
implementation of `Node.get_ranges()` was wrong.

The ranges property is encoded according to the following properties:
 - #address-cells of the bus -> the number of cells for the child
   address
 - #address-cells of the bus's parent -> the number of cells for the
   parent address
 - #size-cells of the bus -> the number of cells for the size

I had also switched the order of the parent address and child address.

All of this is now taken into account with this patch.

Signed-off-by: Nathaniel Graff <nathaniel.graff@sifive.com>